+++ demo/src/main/java/org/vaadin/stefan/ui/view/AbstractCalendarView.java
@@ -31,7 +31,7 @@ public abstract class AbstractCalendarView extends VerticalLayout {
     public AbstractCalendarView() {
         calendar = createCalendar(createDefaultInitialOptions());
 
-        calendar.addThemeVariants(FullCalendarVariant.LUMO);
+        calendar.addThemeVariants(FullCalendarVariant.VAADIN);
 
         calendar.addEntryClickedListener(this::onEntryClick);
         calendar.addEntryDroppedListener(this::onEntryDropped);